refactor(deployments): move MemoryUnit into its own file
fixes https://github.com/openshiftio/openshift.io/issues/2786
