+++ src/app/space/create/deployments/models/scaled-memory-stat.ts
@@ -1,8 +1,8 @@
+import { MemoryStat } from './memory-stat';
 import {
-  MemoryStat,
   MemoryUnit,
   ordinal
-} from './memory-stat';
+} from './memory-unit';
 import { ScaledStat } from './scaled-stat';
 
 import { round } from 'lodash';
