+++ src/app/space/create/deployments/services/deployments.service.ts
@@ -31,11 +31,11 @@ import {
 } from 'lodash';
 
 import { CpuStat } from '../models/cpu-stat';
+import { MemoryStat } from '../models/memory-stat';
 import {
-  MemoryStat,
   MemoryUnit,
   ordinal
-} from '../models/memory-stat';
+} from '../models/memory-unit';
 import { NetworkStat } from '../models/network-stat';
 import { Pods } from '../models/pods';
 import { ScaledMemoryStat } from '../models/scaled-memory-stat';
